fix: 프론트의 요구사항에 따른 Line 생성 시 Section 도 함께 생성, Section.distance 타입 Double -> Integer로 변경

+++ src/main/java/subway/dao/SectionDao.java
@@ -23,7 +23,7 @@ public class SectionDao {
                     rs.getLong("line_id"),
                     rs.getLong("down_station_id"),
                     rs.getLong("up_station_id"),
-                    rs.getDouble("distance")
+                    rs.getInt("distance")
             );
 
     public SectionDao(JdbcTemplate jdbcTemplate, DataSource dataSource) {
